Fix error in start.sh, delete code C in file bison
