+++ parserbison.tab.h
@@ -81,7 +81,7 @@ extern int yydebug;
 #if ! defined YYSTYPE && ! defined YYSTYPE_IS_DECLARED
 union YYSTYPE
 {
-#line 51 "parserbison.y"
+#line 8 "parserbison.y"
 
     int num;    /* para INT */
     char* id;   /* para ID */
