+++ start.sh
@@ -1,5 +1,5 @@
-!/bin/bash
+#!/bin/bash
 bison -d parserbison.y
 flex flex.l
-gcc -o act1 parserbison.tab.c lex.yy.c -lfl
+gcc -o act1 parserbison.tab.c lex.yy.c Tree.c -lfl
 ./act1
